stochastic in heuristic--v1
