+++ src/core/exact_sto_m3ig.hpp
@@ -13,7 +13,9 @@
 
 #ifndef EXACT_STO_M3IG_HPP
 #define EXACT_STO_M3IG_HPP
+#include<vector>
 
+using namespace mockturtle;
 namespace also
 {
   typedef struct Problem_Vector_t_ Problem_Vector_t;
